+++ src/components/providers/modal-provider.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { FC, useEffect, useState } from "react";
-import EmbedModal from "./embed-modal";
-import SubjectListModal from "./subject-list-modal";
-import SubjectViewModal from "./subject-view-modal";
+import EmbedModal from "../modals/embed-modal";
+import SubjectListModal from "../modals/subject-list-modal";
+import SubjectViewModal from "../modals/subject-view-modal";
 
 interface ModalProviderProps {}
