+++ src/components/providers/theme-provider.tsx
@@ -3,7 +3,7 @@
 import { ThemeProvider as NextThemesProvider } from "next-themes";
 import { type ThemeProviderProps } from "next-themes/dist/types";
 import * as React from "react";
-import { ThemeWrapper } from "./theme-wrapper";
+import { ThemeWrapper } from "../theme/theme-wrapper";
 
 export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
     return (
